refactor(save): Move save logic and load logic to separate folders
