+++ code/shared/src/main/scala/maf/save/load/Store.scala
@@ -1,0 +1,6 @@
+package maf.save
+
+import maf.core.Expression
+import maf.modular.AbstractDomain
+
+trait LoadValue[Expr <: Expression] extends Save[Expr] with AbstractDomain[Expr]
